requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==2.2.28
django-allauth==0.41.0
django-widget-tweaks==1.4.12
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==2.0.2
